# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Artalk is a modern comment system based on Golang.",
        "title": "Artalk API",
        "contact": {
            "name": "API Support",
            "url": "https://artalk.js.org",
            "email": "artalkjs@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/ArtalkJS/Artalk/blob/master/LICENSE"
        },
        "version": "2.0"
    },
    "basePath": "/api/v2",
    "paths": {
...
        "/plugins/{plugin_id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get a plugin info by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Plugin"
                ],
                "summary": "Get Plugin Info",
                "operationId": "GetPlugin",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The plugin ID",
                        "name": "plugin_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.ResponsePluginGet"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handler.Map"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handler.Map"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update a plugin status by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Plugin"
                ],
                "summary": "Update Plugin",
                "operationId": "UpdatePlugin",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The plugin ID",
                        "name": "plugin_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "The plugin status",
                        "name": "plugin",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.ParamsPluginUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.ResponsePluginUpdate"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handler.Map"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handler.Map"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/plugins/{plugin_id}/install": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Install a plugin by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Plugin"
                ],
                "summary": "Install Plugin",
                "operationId": "InstallPlugin",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The plugin ID",
                        "name": "plugin_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "The options",
                        "name": "options",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.ParamsPluginInstall"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.Map"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handler.Map"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handler.Map"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "msg": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entity.CookedPlugin": {
            "type": "object",
            "required": [
                "author_link",
                "author_name",
                "compatible",
                "description",
                "donate_link",
                "enabled",
                "id",
                "installed",
                "integrity",
                "local_version",
                "min_artalk_version",
                "name",
                "npm_name",
                "options_schema",
                "repo_link",
                "repo_name",
                "source",
                "type",
                "updated_at",
                "upgrade_available",
                "verified",
                "version"
            ],
            "properties": {
                "author_link": {
                    "type": "string"
                },
                "author_name": {
                    "type": "string"
                },
                "compatible": {
                    "type": "boolean"
                },
                "compatible_notice": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "donate_link": {
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "installed": {
                    "type": "boolean"
                },
                "integrity": {
                    "type": "string"
                },
                "local_version": {
                    "type": "string"
                },
                "min_artalk_version": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "npm_name": {
                    "type": "string"
                },
                "options_schema": {
                    "type": "string"
                },
                "repo_link": {
                    "type": "string"
                },
                "repo_name": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "upgrade_available": {
                    "type": "boolean"
                },
                "verified": {
                    "type": "boolean"
                },
                "version": {
                    "type": "string"
                }
            }
        },
...
        "handler.Map": {
            "type": "object",
            "additionalProperties": true
        },
...
        "handler.ParamsPluginInstall": {
            "type": "object"
        },
...
        "handler.ParamsPluginUpdate": {
            "type": "object",
            "required": [
                "enabled"
            ],
            "properties": {
                "client_options": {
                    "description": "The plugin client options (JSON string)",
                    "type": "string"
                },
                "enabled": {
                    "description": "The plugin enabled status",
                    "type": "boolean"
                }
            }
        },
...
        "handler.ResponsePluginGet": {
            "type": "object",
            "required": [
                "client_options",
                "enabled",
                "options_schema",
                "plugin"
            ],
            "properties": {
                "client_options": {
                    "description": "The plugin client options (JSON string)",
                    "type": "string"
                },
                "enabled": {
                    "description": "The plugin enabled status",
                    "type": "boolean"
                },
                "options_schema": {
                    "description": "The plugin options schema (JSON string)",
                    "type": "string"
                },
                "plugin": {
                    "description": "The plugin info",
                    "allOf": [
                        {
                            "$ref": "#/definitions/entity.CookedPlugin"
                        }
                    ]
                }
            }
        },
...
        "handler.ResponsePluginUpdate": {
            "type": "object",
            "required": [
                "plugin"
            ],
            "properties": {
                "plugin": {
                    "$ref": "#/definitions/entity.CookedPlugin"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "\"Type 'Bearer TOKEN' to correctly set the API Key\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}